refactor(db): split db/schedules.py into a db/schedules/ mixin package (#1481)
Pure mechanical move of the 4,139-line src/backend/db/schedules.py into a
ten-mixin db/schedules/ package composed into ScheduleOperations — the
sanctioned db/agent_settings/ shape (Invariant #2). Zero behavior change,
zero schema change (0 DDL in the file, Invariant #3 untouched), zero WHERE
clause edits, zero signature edits.

- __init__.py composes ScheduleOperations from ScheduleCommonMixin / Crud /
  Webhooks / Executions / Queue / Cleanup / Analytics / Stats / GitConfig /
  Retention and re-exports the class + _norm_ts + _TRIGGER_BUCKETS so the
  facade import `from db.schedules import ScheduleOperations` is unchanged
  (both prod sites: database.py, db/__init__.py). No import edges between
  mixin files; cross-slice refs resolve via the composed class's MRO.
- find_soft_deleted_schedules_past_retention moves to retention.py so its
  caller, count_soft_deleted_schedules_past_retention, and the shared
  _soft_deleted_schedules_predicate module-global are co-located (a bare-name
  module-global does not resolve via MRO — would NameError on the #834 purge).
- The three fat signatures (create_task_execution / update_execution_status /
  create_schedule_execution) are byte-identical (reserved for #1482).
- The 20 #1082 status-CAS writers and all pull/backlog/lease seams (#1081/#1550,
  BACKLOG-001) moved verbatim; every WHERE precondition preserved.
- Only allowlisted text deviations: relative-import depth (.->..), logger name
  pinned to the literal getLogger("db.schedules") in the 3 logging modules, and
  the _PERCENTILE_ROWSET_CAP monkeypatch doc-comment retargeted to
  db.schedules.analytics.

Lockstep test edits forced by the move (the only non-move work):
- test_schedule_status_observability.py: writer discovery now rglob-scans the
  db/schedules/ package (was one hardcoded file) with an in-function
  `assert sites` non-empty tripwire so the CAS-precondition guard can never pass
  vacuously; found == _EXPECTED_UPDATE_SITES (20) both ways.
- test_agent_analytics.py / test_schedule_analytics.py: _PERCENTILE_ROWSET_CAP
  monkeypatch retargeted to db.schedules.analytics (module-identity trap).
- conftest.py + test_agent_analytics.py: sys.modules eviction/baseline widened
  to db.schedules.* children (prefix match, future-sub-split-proof).

Proof: symbol-set diff (103 functions, none added/dropped/renamed) + AST
source-segment body diff (all 103 bodies byte-identical) vs origin/dev.
